requirements_conda.txt
numpy
scipy>=1.9.0
matplotlib
pandas
opencv
networkx
plyfile
pytorch>=1.8.0
torchvision
optuna